fix: migrate db to mysql using flyway refactor : all repositories & tables with new schema design

+++ backend/src/main/java/com/smalltrend/repository/PermissionRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 @Repository
-public interface PermissionRepository extends JpaRepository<Permission, Long> {
+public interface PermissionRepository extends JpaRepository<Permission, Integer> {
+
     Optional<Permission> findByName(String name);
 }
