+++ backend/src/main/java/com/smalltrend/repository/RolePermissionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface RolePermissionRepository extends JpaRepository<RolePermission, Long> {
+public interface RolePermissionRepository extends JpaRepository<RolePermission, Integer> {
 }
